Change Int128 to Decimal

+++ compiler/const/conrex.vb
@@ -1,6 +1,6 @@
 ﻿Public Class conrex
     Public Shared yocommondatatype() As String = {"bool", "i8", "i16", "i32", "i64", "i128", "ui8", "ui16", "ui32", "u64", "str", "chr", "f32", "f64"}
-    Public Shared msilcommondatatype() As String = {"Boolean", "SByte", "Int16", "Int32", "Int64", "Int128", "Byte", "UInt16", "UInt32", "UInt64", "String", "Char", "Single", "Double"}
+    Public Shared msilcommondatatype() As String = {"Boolean", "SByte", "Int16", "Int32", "Int64", "Decimal", "Byte", "UInt16", "UInt32", "UInt64", "String", "Char", "Single", "Double"}
     Public Shared specificxmlchar() As Char = {">", "<", "'", """", "&"}
     Private Shared rand As New Random
     Public Shared specificrandomnumber As Integer = rand.Next(10000, 99999)
